# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Nepali Date Converter API. Convert between Bikram Sambat (BS) and Anno Domini (AD), get today's Nepali date, list calendar months, festivals, and fiscal year mappings.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface NepaliDate {
  year: number;
  month: number;
  day: number;
  monthName: string;
  monthNameNepali: string;
  /** 0 = Sunday, 6 = Saturday */
  weekday: number;
  weekdayName: string;
  weekdayNameNepali: string;
  /** Human readable e.g. "Baisakh 15, 2082" */
  formatted: string;
  formattedNepali: string;
}

export interface GregorianDate {
  year: number;
  month: number;
  day: number;
  monthName: string;
  weekday: number;
  weekdayName: string;
  /** ISO date string YYYY-MM-DD */
  iso: string;
  formatted: string;
}

export interface ConversionResult {
  bs: NepaliDate;
  ad: GregorianDate;
}

export interface CalendarDay {
  bsYear: number;
  bsMonth: number;
  bsDay: number;
  adIso: string;
  weekday: number;
  isHoliday: boolean;
  /** @nullable */
  festival: string | null;
}

export interface CalendarMonth {
  year: number;
  month: number;
  monthName: string;
  monthNameNepali: string;
  daysInMonth: number;
  startWeekday: number;
  days: CalendarDay[];
}

export interface Festival {
  name: string;
  nameNepali: string;
  bsYear: number;
  bsMonth: number;
  bsDay: number;
  adIso: string;
  /** e.g. "Religious", "National", "Regional" */
  category: string;
  isPublicHoliday: boolean;
}

export interface FiscalYear {
  /** e.g. "FY 2082/83" */
  label: string;
  startBs: NepaliDate;
  endBs: NepaliDate;
  startAd: GregorianDate;
  endAd: GregorianDate;
}

export type ConvertBsToAdParams = {
  /**
   * @minimum 1970
   * @maximum 2100
   */
  year: number;
  /**
   * @minimum 1
   * @maximum 12
   */
  month: number;
  /**
   * @minimum 1
   * @maximum 32
   */
  day: number;
};

export type ConvertAdToBsParams = {
  /**
   * @minimum 1913
   * @maximum 2043
   */
  year: number;
  /**
   * @minimum 1
   * @maximum 12
   */
  month: number;
  /**
   * @minimum 1
   * @maximum 31
   */
  day: number;
};