# lib/api-zod/src/generated/types/calendarMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Nepali Date Converter API. Convert between Bikram Sambat (BS) and Anno Domini (AD), get today's Nepali date, list calendar months, festivals, and fiscal year mappings.
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarDay } from "./calendarDay";

export interface CalendarMonth {
  year: number;
  month: number;
  monthName: string;
  monthNameNepali: string;
  daysInMonth: number;
  startWeekday: number;
  days: CalendarDay[];
}