# lib/api-zod/src/generated/types/festival.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Nepali Date Converter API. Convert between Bikram Sambat (BS) and Anno Domini (AD), get today's Nepali date, list calendar months, festivals, and fiscal year mappings.
 * OpenAPI spec version: 0.1.0
 */

export interface Festival {
  name: string;
  nameNepali: string;
  bsYear: number;
  bsMonth: number;
  bsDay: number;
  adIso: string;
  /** e.g. "Religious", "National", "Regional" */
  category: string;
  isPublicHoliday: boolean;
}