# lib/api-zod/src/generated/types/fiscalYear.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Nepali Date Converter API. Convert between Bikram Sambat (BS) and Anno Domini (AD), get today's Nepali date, list calendar months, festivals, and fiscal year mappings.
 * OpenAPI spec version: 0.1.0
 */
import type { GregorianDate } from "./gregorianDate";
import type { NepaliDate } from "./nepaliDate";

export interface FiscalYear {
  /** e.g. "FY 2082/83" */
  label: string;
  startBs: NepaliDate;
  endBs: NepaliDate;
  startAd: GregorianDate;
  endAd: GregorianDate;
}