# lib/api-zod/src/generated/types/nepaliDate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Nepali Date Converter API. Convert between Bikram Sambat (BS) and Anno Domini (AD), get today's Nepali date, list calendar months, festivals, and fiscal year mappings.
 * OpenAPI spec version: 0.1.0
 */

export interface NepaliDate {
  year: number;
  month: number;
  day: number;
  monthName: string;
  monthNameNepali: string;
  /** 0 = Sunday, 6 = Saturday */
  weekday: number;
  weekdayName: string;
  weekdayNameNepali: string;
  /** Human readable e.g. "Baisakh 15, 2082" */
  formatted: string;
  formattedNepali: string;
}